refactoring code and simplifying Dockerfile
